translations

+++ src/operations/channel/deleteChannel.ts
@@ -1,9 +1,14 @@
 import db from '../../db/db.js';
 import { ModelType, MutationType } from '../../enums.js';
+import libData from '../../helpers/libData.js';
 import { Channel } from '../../types/models/Channel.js';
 import { MutationResult } from '../../types/MutationResult.js';
 
 const deleteChannel = async (id: string): Promise<MutationResult<Channel>> => {
+  if (!libData.isInitialized()) {
+    throw new Error('not-initialized');
+  }
+
   try {
     await db.delete(id, ModelType.Channel);
