+++ src/operations/channelMessage/deleteChannelMessage.ts
@@ -1,11 +1,16 @@
 import db from '../../db/db.js';
 import { ModelType, MutationType } from '../../enums.js';
+import libData from '../../helpers/libData.js';
 import { ChannelMessage } from '../../types/models/ChannelMessage.js';
 import { MutationResult } from '../../types/MutationResult.js';
 
 const deleteChannelMessage = async (
   id: string,
 ): Promise<MutationResult<ChannelMessage>> => {
+  if (!libData.isInitialized()) {
+    throw new Error('not-initialized');
+  }
+
   try {
     await db.delete(id, ModelType.ChannelMessage);
