+++ src/operations/channelParticipant/deleteChannelParticipant.ts
@@ -1,11 +1,16 @@
 import db from '../../db/db.js';
 import { ModelType, MutationType } from '../../enums.js';
+import libData from '../../helpers/libData.js';
 import { ChannelParticipant } from '../../types/models/ChannelParticipant.js';
 import { MutationResult } from '../../types/MutationResult.js';
 
 const deleteChannelParticipant = async (
   id: string,
 ): Promise<MutationResult<ChannelParticipant>> => {
+  if (!libData.isInitialized()) {
+    throw new Error('not-initialized');
+  }
+
   try {
     await db.delete(id, ModelType.ChannelParticipant);
